REFACTOR: moved mocks to service package
